+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## 1.3.5
+## 1.3.6
 
 ### Bug Fixes